Fix up stuff from merge (reduce diff to branch)

+++ reactos/ntoskrnl/mm/ARM3/contmem.c
@@ -636,7 +636,7 @@ MmFreeContiguousMemory(IN PVOID BaseAddress)
 VOID
 NTAPI
 MmFreeContiguousMemorySpecifyCache(IN PVOID BaseAddress,
-                                   IN ULONG NumberOfBytes,
+                                   IN SIZE_T NumberOfBytes,
                                    IN MEMORY_CACHING_TYPE CacheType)
 {
     //
